Fix unnecessary home directory creation issues
