+++ protocol/cmd/dydxprotocold/main.go
@@ -14,12 +14,7 @@ func main() {
 	config.SetupConfig()
 
 	option := cmd.GetOptionWithCustomStartCmd()
-	// [ Home Dir Temp Fix ] (also see protocol/cmd/dydxprotocold/cmd/root.go)
-	// We pass in a tempdir as a temporary hack until fixed in cosmos.
-	// If not, and we pass in a custom home dir, it will try read or create the default home dir
-	// before using the custom home dir.
-	// See https://github.com/cosmos/cosmos-sdk/issues/18868.
-	rootCmd := cmd.NewRootCmd(option, tempDir())
+	rootCmd := cmd.NewRootCmd(option, app.DefaultNodeHome)
 
 	cmd.AddTendermintSubcommands(rootCmd)
 	cmd.AddInitCmdPostRunE(rootCmd)
@@ -27,14 +22,4 @@ func main() {
 	if err := svrcmd.Execute(rootCmd, constants.AppDaemonName, app.DefaultNodeHome); err != nil {
 		os.Exit(1)
 	}
-}
-
-var tempDir = func() string {
-	dir, err := os.MkdirTemp("", "dydxprotocol")
-	if err != nil {
-		dir = app.DefaultNodeHome
-	}
-	defer os.RemoveAll(dir)
-
-	return dir
 }
